+++ src/itemCheckbox/ItemCheckbox.tsx
@@ -71,6 +71,11 @@ class ItemCheckbox extends Component<ItemCheckboxProps> {
         rightTitle={data}
         rightTitleDisplay={TextDisplayType.SUBHEADERSTRONG}
         rightBody={dataInfo}
+        /* No a11y issue here
+          - The input is well wrapped with the label
+          - The linter can't access the complex components implementation
+        */
+        // eslint-disable-next-line jsx-a11y/label-has-associated-control
         tag={<label />}
         rightAddon={checkbox}
         isClickable